minor fixes

+++ WickedEngine/wiGraphicsDevice_DX11.cpp
@@ -3208,7 +3208,7 @@ void GraphicsDevice_DX11::MSAAResolve(Texture2D* pDst, const Texture2D* pSrc, GR
 void GraphicsDevice_DX11::UpdateBuffer(GPUBuffer* buffer, const void* data, GRAPHICSTHREAD threadID, int dataSize)
 {
 	assert(buffer->desc.Usage != USAGE_IMMUTABLE && "Cannot update IMMUTABLE GPUBuffer!");
-	assert(buffer->desc.ByteWidth >= dataSize || dataSize < 0 && "Data size is too big!");
+	assert((int)buffer->desc.ByteWidth >= dataSize || dataSize < 0 && "Data size is too big!");
 
 	if (dataSize == 0)
 	{
